Add support for fractional/decimal quantities in medicine inventory
Features:
- Support selling fractions like 0.5 ml, 1.5 ml, 2.75 ml
- Essential for cosmetics, fillers, and injectable products
- Allows up to 2 decimal places (e.g., 0.5, 1.5, 10.25)

Database Changes:
- Changed quantity columns from integer to decimal(10,2):
  * medicines.stock_quantity
  * medicine_transactions.quantity
  * medicine_transactions.stock_before
  * medicine_transactions.stock_after
- All existing data preserved during migration

Model Updates:
- Medicine model: stock_quantity cast as decimal:2
- MedicineTransaction model: all quantity fields cast as decimal:2
- Supports fractional calculations throughout system

Validation Updates:
- MedicineController::store() - Accept decimal stock quantities
- MedicineController::update() - Accept decimal stock quantities
- MedicineController::processSell() - Accept decimal sale quantities
- MedicineController::processPurchase() - Accept decimal purchase quantities
- SimplePrescriptionController::convertToSale() - Support decimal dispensing
- Regex validation: /^\d+(\.\d{1,2})?$/ (max 2 decimal places)
- Custom error messages for invalid decimal formats

View Updates:
- Sell form: input step='0.01', min='0.01'
- Purchase form: input step='0.01', min='0.01'
- Helper text: 'Supports decimals (e.g., 0.5, 1.5, 10.25)'
- Shows max available stock with decimal support

Use Cases:
- Dermal fillers: Sell 0.5 ml, 1.0 ml, 1.5 ml
- Botox injections: Sell 2.5 units, 5.5 units
- Cosmetic creams: Sell 0.25 ml samples
- Liquid medications: Sell 7.5 ml doses
- Any product requiring fractional quantities

Examples:
- Enter 0.5 in quantity → Sells half unit
- Enter 1.5 in quantity → Sells one and a half units
- Enter 10.25 in quantity → Sells ten and quarter units
- Stock: 50.75 units → Displayed with 2 decimals

Technical Details:
- decimal(10,2) = Max 99,999,999.99 units
- Calculations maintain precision
- No rounding errors
- Full audit trail with decimals
- Compatible with existing integer data

+++ app/Http/Controllers/SimplePrescriptionController.php
@@ -721,8 +721,8 @@ class SimplePrescriptionController extends Controller
                     continue;
                 }
 
-                // Check stock quantity
-                $quantityNeeded = (int) ($prescribedMedicine->quantity ?? 1);
+                // Check stock quantity - support decimals for cosmetics/fillers
+                $quantityNeeded = (float) ($prescribedMedicine->quantity ?? 1);
                 if ($medicine->stock_quantity < $quantityNeeded) {
                     $insufficientStock[] = [
                         'name' => $medicine->name,
@@ -761,7 +761,8 @@ class SimplePrescriptionController extends Controller
                     ->where('is_active', true)
                     ->first();
 
-                $quantityNeeded = (int) ($prescribedMedicine->quantity ?? 1);
+                // Support decimal quantities for cosmetics/fillers (e.g., 0.5 ml, 1.5 ml)
+                $quantityNeeded = (float) ($prescribedMedicine->quantity ?? 1);
                 $unitPrice = $medicine->selling_price ?? 0;
                 $itemTotal = $quantityNeeded * $unitPrice;
                 $totalAmount += $itemTotal;
